requirements.txt
gradio
pandas
pdfplumber
faiss-cpu
sentence-transformers
numpy
